fix: change constructServiceURL from method to function
Now it can be used as a utility function instead of an inherited method.

An SDK class can now include a `constructServiceURL` method with a different signature than this function.
